refactor: классы перенесены в отдельный пакет
