--- hw11-cache/src/main/java/ru/dankoy/otus/hibernate/cache/core/dao/AddressDataSetDao.java
@@ -1,18 +1,0 @@
-package ru.dankoy.otus.hibernate.cache.core.dao;
-
-import ru.dankoy.otus.hibernate.cache.core.model.AddressDataSet;
-import ru.dankoy.otus.hibernate.cache.core.sessionmanager.SessionManager;
-
-import java.util.Optional;
-
-public interface AddressDataSetDao {
-    Optional<AddressDataSet> findById(long id);
-
-    long insertAddress(AddressDataSet addressDataSet);
-
-    void updateAddress(AddressDataSet addressDataSet);
-
-    void insertOrUpdate(AddressDataSet addressDataSet);
-
-    SessionManager getSessionManager();
-}
