--- hw11-cache/src/main/java/ru/dankoy/otus/hibernate/cache/core/dao/AddressDataSetDaoException.java
@@ -1,7 +1,0 @@
-package ru.dankoy.otus.hibernate.cache.core.dao;
-
-public class AddressDataSetDaoException extends RuntimeException {
-    public AddressDataSetDaoException(Exception ex) {
-        super(ex);
-    }
-}
