--- hw11-cache/src/main/java/ru/dankoy/otus/hibernate/cache/core/dao/PhoneDataSetDao.java
@@ -1,19 +1,0 @@
-package ru.dankoy.otus.hibernate.cache.core.dao;
-
-import ru.dankoy.otus.hibernate.cache.core.model.PhoneDataSet;
-import ru.dankoy.otus.hibernate.cache.core.sessionmanager.SessionManager;
-
-import java.util.Optional;
-
-public interface PhoneDataSetDao {
-    Optional<PhoneDataSet> findById(long id);
-
-    long insertPhone(PhoneDataSet phoneDataSet);
-
-    void updatePhone(PhoneDataSet phoneDataSet);
-
-    void insertOrUpdate(PhoneDataSet phoneDataSet);
-
-    SessionManager getSessionManager();
-
-}
