--- hw11-cache/src/main/java/ru/dankoy/otus/hibernate/cache/core/dao/PhoneDataSetDaoException.java
@@ -1,7 +1,0 @@
-package ru.dankoy.otus.hibernate.cache.core.dao;
-
-public class PhoneDataSetDaoException extends RuntimeException {
-    public PhoneDataSetDaoException(Exception ex) {
-        super(ex);
-    }
-}
