--- hw11-cache/src/main/java/ru/dankoy/otus/hibernate/cache/core/dao/UserDao.java
@@ -1,18 +1,0 @@
-package ru.dankoy.otus.hibernate.cache.core.dao;
-
-import ru.dankoy.otus.hibernate.cache.core.model.User;
-import ru.dankoy.otus.hibernate.cache.core.sessionmanager.SessionManager;
-
-import java.util.Optional;
-
-public interface UserDao {
-    Optional<User> findById(long id);
-
-    long insertUser(User user);
-
-    void updateUser(User user);
-
-    void insertOrUpdate(User user);
-
-    SessionManager getSessionManager();
-}
