--- hw11-cache/src/main/java/ru/dankoy/otus/hibernate/cache/core/dao/UserDaoException.java
@@ -1,7 +1,0 @@
-package ru.dankoy.otus.hibernate.cache.core.dao;
-
-public class UserDaoException extends RuntimeException {
-    public UserDaoException(Exception ex) {
-        super(ex);
-    }
-}
