--- hw11-cache/src/main/java/ru/dankoy/otus/hibernate/cache/core/service/DbServiceException.java
@@ -1,7 +1,0 @@
-package ru.dankoy.otus.hibernate.cache.core.service;
-
-public class DbServiceException extends RuntimeException {
-    public DbServiceException(Exception e) {
-        super(e);
-    }
-}
