--- hw11-cache/src/main/java/ru/dankoy/otus/hibernate/cache/core/sessionmanager/DatabaseSession.java
@@ -1,4 +1,0 @@
-package ru.dankoy.otus.hibernate.cache.core.sessionmanager;
-
-public interface DatabaseSession {
-}
