--- hw11-cache/src/main/java/ru/dankoy/otus/hibernate/cache/core/sessionmanager/SessionManager.java
@@ -1,13 +1,0 @@
-package ru.dankoy.otus.hibernate.cache.core.sessionmanager;
-
-public interface SessionManager extends AutoCloseable {
-    void beginSession();
-
-    void commitSession();
-
-    void rollbackSession();
-
-    void close();
-
-    DatabaseSession getCurrentSession();
-}
