--- hw11-cache/src/main/java/ru/dankoy/otus/hibernate/cache/core/sessionmanager/SessionManagerException.java
@@ -1,12 +1,0 @@
-package ru.dankoy.otus.hibernate.cache.core.sessionmanager;
-
-
-public class SessionManagerException extends RuntimeException {
-    public SessionManagerException(String msg) {
-        super(msg);
-    }
-
-    public SessionManagerException(Exception ex) {
-        super(ex);
-    }
-}
