--- hw11-cache/src/main/java/ru/dankoy/otus/hibernate/cache/flyway/MigrationsExecutor.java
@@ -1,8 +1,0 @@
-package ru.dankoy.otus.hibernate.cache.flyway;
-
-public interface MigrationsExecutor {
-
-    void cleanDb();
-    void executeMigrations();
-
-}
