--- hw11-cache/src/main/java/ru/dankoy/otus/hibernate/cache/h2/DataSourceH2.java
@@ -1,64 +1,0 @@
-package ru.dankoy.otus.hibernate.cache.h2;
-
-import javax.sql.DataSource;
-import java.io.PrintWriter;
-import java.sql.Connection;
-import java.sql.DriverManager;
-import java.sql.SQLException;
-import java.util.logging.Logger;
-
-/**
- * @author sergey
- * created on 03.02.19.
- */
-public class DataSourceH2 implements DataSource {
-    private static final String URL = "jdbc:h2:~/test;DB_CLOSE_DELAY=-1;DB_CLOSE_ON_EXIT=FALSE;AUTO_SERVER=TRUE";
-
-    @Override
-    public Connection getConnection() throws SQLException {
-        var connection = DriverManager.getConnection(URL);
-        connection.setAutoCommit(false);
-        return connection;
-    }
-
-    @Override
-    public Connection getConnection(String username, String password) throws SQLException {
-        return getConnection();
-    }
-
-    @Override
-    public PrintWriter getLogWriter() {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public void setLogWriter(PrintWriter out) {
-        throw new UnsupportedOperationException();
-
-    }
-
-    @Override
-    public int getLoginTimeout() {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public void setLoginTimeout(int seconds) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public Logger getParentLogger() {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public <T> T unwrap(Class<T> iface) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public boolean isWrapperFor(Class<?> iface) {
-        throw new UnsupportedOperationException();
-    }
-}
