--- hw11-cache/src/main/java/ru/dankoy/otus/hibernate/cache/hibernate/sessionmanager/DatabaseSessionHibernate.java
@@ -1,31 +1,0 @@
-package ru.dankoy.otus.hibernate.cache.hibernate.sessionmanager;
-
-import org.hibernate.Session;
-import org.hibernate.Transaction;
-import ru.dankoy.otus.hibernate.cache.core.sessionmanager.DatabaseSession;
-
-
-public class DatabaseSessionHibernate implements DatabaseSession {
-    private final Session session;
-    private final Transaction transaction;
-
-    DatabaseSessionHibernate(Session session) {
-        this.session = session;
-        this.transaction = session.beginTransaction();
-    }
-
-    public Session getHibernateSession() {
-        return session;
-    }
-
-    public Transaction getTransaction() {
-        return transaction;
-    }
-
-    public void close() {
-        if (transaction.isActive()) {
-            transaction.commit();
-        }
-        session.close();
-    }
-}
