--- hw11-cache/src/main/java/ru/dankoy/otus/hibernate/cache/jdbc/DbExecutor.java
@@ -1,15 +1,0 @@
-package ru.dankoy.otus.hibernate.cache.jdbc;
-
-import java.sql.Connection;
-import java.sql.ResultSet;
-import java.sql.SQLException;
-import java.util.List;
-import java.util.Optional;
-import java.util.function.Function;
-
-public interface DbExecutor<T> {
-
-    long executeInsert(Connection connection, String sql, List<Object> params) throws SQLException;
-
-    Optional<T> executeSelect(Connection connection, String sql, Object id, Function<ResultSet, T> rsHandler) throws SQLException;
-}
