--- hw11-cache/src/main/java/ru/dankoy/otus/hibernate/cache/jdbc/sessionmanager/DatabaseSessionJdbc.java
@@ -1,17 +1,0 @@
-package ru.dankoy.otus.hibernate.cache.jdbc.sessionmanager;
-
-import ru.dankoy.otus.hibernate.cache.core.sessionmanager.DatabaseSession;
-
-import java.sql.Connection;
-
-public class DatabaseSessionJdbc implements DatabaseSession {
-    private final Connection connection;
-
-    DatabaseSessionJdbc(Connection connection) {
-        this.connection = connection;
-    }
-
-    public Connection getConnection() {
-        return connection;
-    }
-}
