+++ hw11-cache/src/main/java/ru/dankoy/otus/hibernate/core/dao/AddressDataSetDaoException.java
@@ -1,0 +1,7 @@
+package ru.dankoy.otus.hibernate.core.dao;
+
+public class AddressDataSetDaoException extends RuntimeException {
+    public AddressDataSetDaoException(Exception ex) {
+        super(ex);
+    }
+}
