+++ hw11-cache/src/main/java/ru/dankoy/otus/hibernate/core/dao/PhoneDataSetDao.java
@@ -1,0 +1,19 @@
+package ru.dankoy.otus.hibernate.core.dao;
+
+import ru.dankoy.otus.hibernate.core.model.PhoneDataSet;
+import ru.dankoy.otus.hibernate.core.sessionmanager.SessionManager;
+
+import java.util.Optional;
+
+public interface PhoneDataSetDao {
+    Optional<PhoneDataSet> findById(long id);
+
+    long insertPhone(PhoneDataSet phoneDataSet);
+
+    void updatePhone(PhoneDataSet phoneDataSet);
+
+    void insertOrUpdate(PhoneDataSet phoneDataSet);
+
+    SessionManager getSessionManager();
+
+}
