+++ hw11-cache/src/main/java/ru/dankoy/otus/hibernate/core/dao/PhoneDataSetDaoException.java
@@ -1,0 +1,7 @@
+package ru.dankoy.otus.hibernate.core.dao;
+
+public class PhoneDataSetDaoException extends RuntimeException {
+    public PhoneDataSetDaoException(Exception ex) {
+        super(ex);
+    }
+}
