+++ hw11-cache/src/main/java/ru/dankoy/otus/hibernate/core/dao/UserDao.java
@@ -1,0 +1,18 @@
+package ru.dankoy.otus.hibernate.core.dao;
+
+import ru.dankoy.otus.hibernate.core.model.User;
+import ru.dankoy.otus.hibernate.core.sessionmanager.SessionManager;
+
+import java.util.Optional;
+
+public interface UserDao {
+    Optional<User> findById(long id);
+
+    long insertUser(User user);
+
+    void updateUser(User user);
+
+    void insertOrUpdate(User user);
+
+    SessionManager getSessionManager();
+}
