+++ hw11-cache/src/main/java/ru/dankoy/otus/hibernate/core/dao/UserDaoException.java
@@ -1,0 +1,7 @@
+package ru.dankoy.otus.hibernate.core.dao;
+
+public class UserDaoException extends RuntimeException {
+    public UserDaoException(Exception ex) {
+        super(ex);
+    }
+}
