+++ hw11-cache/src/main/java/ru/dankoy/otus/hibernate/core/service/DbServiceException.java
@@ -1,0 +1,7 @@
+package ru.dankoy.otus.hibernate.core.service;
+
+public class DbServiceException extends RuntimeException {
+    public DbServiceException(Exception e) {
+        super(e);
+    }
+}
