+++ hw11-cache/src/main/java/ru/dankoy/otus/hibernate/core/service/userservice/DbServiceUserImpl.java
@@ -1,0 +1,53 @@
+package ru.dankoy.otus.hibernate.core.service.userservice;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import ru.dankoy.otus.hibernate.core.dao.UserDao;
+import ru.dankoy.otus.hibernate.core.model.User;
+import ru.dankoy.otus.hibernate.core.service.DbServiceException;
+
+import java.util.Optional;
+
+public class DbServiceUserImpl implements DBServiceUser {
+    private static final Logger logger = LoggerFactory.getLogger(DbServiceUserImpl.class);
+
+    private final UserDao userDao;
+
+    public DbServiceUserImpl(UserDao userDao) {
+        this.userDao = userDao;
+    }
+
+    @Override
+    public long saveUser(User user) {
+        try (var sessionManager = userDao.getSessionManager()) {
+            sessionManager.beginSession();
+            try {
+                var userId = userDao.insertUser(user);
+                sessionManager.commitSession();
+
+                logger.info("created user: {}", userId);
+                return userId;
+            } catch (Exception e) {
+                sessionManager.rollbackSession();
+                throw new DbServiceException(e);
+            }
+        }
+    }
+
+    @Override
+    public Optional<User> getUser(long id) {
+        try (var sessionManager = userDao.getSessionManager()) {
+            sessionManager.beginSession();
+            try {
+                Optional<User> userOptional = userDao.findById(id);
+
+                logger.info("user: {}", userOptional.orElse(null));
+                return userOptional;
+            } catch (Exception e) {
+                logger.error(e.getMessage(), e);
+                sessionManager.rollbackSession();
+            }
+            return Optional.empty();
+        }
+    }
+}
