+++ hw11-cache/src/main/java/ru/dankoy/otus/hibernate/core/sessionmanager/DatabaseSession.java
@@ -1,0 +1,4 @@
+package ru.dankoy.otus.hibernate.core.sessionmanager;
+
+public interface DatabaseSession {
+}
