+++ hw11-cache/src/main/java/ru/dankoy/otus/hibernate/core/sessionmanager/SessionManager.java
@@ -1,0 +1,13 @@
+package ru.dankoy.otus.hibernate.core.sessionmanager;
+
+public interface SessionManager extends AutoCloseable {
+    void beginSession();
+
+    void commitSession();
+
+    void rollbackSession();
+
+    void close();
+
+    DatabaseSession getCurrentSession();
+}
