+++ hw11-cache/src/main/java/ru/dankoy/otus/hibernate/core/sessionmanager/SessionManagerException.java
@@ -1,0 +1,12 @@
+package ru.dankoy.otus.hibernate.core.sessionmanager;
+
+
+public class SessionManagerException extends RuntimeException {
+    public SessionManagerException(String msg) {
+        super(msg);
+    }
+
+    public SessionManagerException(Exception ex) {
+        super(ex);
+    }
+}
