+++ hw11-cache/src/main/java/ru/dankoy/otus/hibernate/flyway/MigrationsExecutor.java
@@ -1,0 +1,8 @@
+package ru.dankoy.otus.hibernate.flyway;
+
+public interface MigrationsExecutor {
+
+    void cleanDb();
+    void executeMigrations();
+
+}
