+++ hw11-cache/src/main/java/ru/dankoy/otus/hibernate/hibernate/dao/UserDaoHibernate.java
@@ -1,0 +1,80 @@
+package ru.dankoy.otus.hibernate.hibernate.dao;
+
+import org.hibernate.Session;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import ru.dankoy.otus.hibernate.core.dao.UserDao;
+import ru.dankoy.otus.hibernate.core.dao.UserDaoException;
+import ru.dankoy.otus.hibernate.core.model.User;
+import ru.dankoy.otus.hibernate.core.sessionmanager.SessionManager;
+import ru.dankoy.otus.hibernate.hibernate.sessionmanager.DatabaseSessionHibernate;
+import ru.dankoy.otus.hibernate.hibernate.sessionmanager.SessionManagerHibernate;
+
+import java.util.Optional;
+
+public class UserDaoHibernate implements UserDao {
+
+    private static Logger logger = LoggerFactory.getLogger(UserDaoHibernate.class);
+
+    private final SessionManagerHibernate sessionManager;
+
+    public UserDaoHibernate(SessionManagerHibernate sessionManager) {
+        this.sessionManager = sessionManager;
+    }
+
+
+    @Override
+    public Optional<User> findById(long id) {
+        DatabaseSessionHibernate currentSession = sessionManager.getCurrentSession();
+        try {
+            return Optional.ofNullable(currentSession.getHibernateSession().find(User.class, id));
+        } catch (Exception e) {
+            logger.error(e.getMessage(), e);
+        }
+        return Optional.empty();
+    }
+
+    @Override
+    public long insertUser(User user) {
+        DatabaseSessionHibernate currentSession = sessionManager.getCurrentSession();
+        try {
+            Session hibernateSession = currentSession.getHibernateSession();
+            hibernateSession.persist(user);
+            hibernateSession.flush();
+            return user.getId();
+        } catch (Exception e) {
+            logger.error(e.getMessage(), e);
+            throw new UserDaoException(e);
+        }
+    }
+
+    @Override
+    public void updateUser(User user) {
+        DatabaseSessionHibernate currentSession = sessionManager.getCurrentSession();
+        try {
+            Session hibernateSession = currentSession.getHibernateSession();
+            hibernateSession.merge(user);
+        } catch (Exception e) {
+            logger.error(e.getMessage(), e);
+            throw new UserDaoException(e);
+        }
+    }
+
+    @Override
+    public void insertOrUpdate(User user) {
+        DatabaseSessionHibernate currentSession = sessionManager.getCurrentSession();
+        try {
+            Session hibernateSession = currentSession.getHibernateSession();
+            hibernateSession.persist(user);
+            hibernateSession.flush();
+        } catch (Exception e) {
+            logger.error(e.getMessage(), e);
+            throw new UserDaoException(e);
+        }
+    }
+
+    @Override
+    public SessionManager getSessionManager() {
+        return sessionManager;
+    }
+}
